mostly moving cost_power_law_flow
